requirements.txt
appdirs==1.4.3
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bandit==1.6.2
black==19.3b0
bleach==4.0.0
cffi==1.14.6
Click==7.0
colorlover==0.3.0
coverage==4.5.4
cufflinks==0.17.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.7.1
Django==2.2.4
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
djangorestframework==3.10.2
entrypoints==0.3
gitdb2==2.0.5
GitPython==3.0.2
gunicorn==19.9.0
importlib-metadata==0.19
ipdb==0.12.2
ipykernel==5.5.5
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.15.1
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.2
numpy==1.19.5
opencv-python==4.5.3.56
packaging==19.1
pandas==1.1.5
pandocfilters==1.4.3
parso==0.5.1
pbr==5.4.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.3.1
pip-tools==4.1.0
plotly==5.1.0
pluggy==0.12.0
prometheus-client==0.11.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.4.2
pylama==7.7.1
pyparsing==2.4.2
pyrsistent==0.18.0
pytest==5.1.1
pytest-bandit==0.5.2
pytest-cov==2.7.1
pytest-django==3.5.1
python-dateutil==2.8.2
pytz==2019.2
PyYAML==5.1.2
pyzmq==22.2.1
scipy==1.5.4
Send2Trash==1.7.1
six==1.12.0
smmap2==2.0.5
snowballstemmer==1.9.0
sqlparse==0.3.0
stevedore==1.30.1
tenacity==8.0.1
terminado==0.10.1
testpath==0.5.0
toml==0.10.0
tornado==6.1
traitlets==4.3.2
typing-extensions==3.10.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
zipp==0.6.0
